djangoReactProject/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
boto3==1.34.21
botocore==1.34.21
cachetools==5.3.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
fire==0.5.0
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.20.2
idna==2.10
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
mysqlclient==2.2.1
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.23.4
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.16
rsa==4.9
s3transfer==0.10.0
sahi==0.11.15
scipy==1.11.4
seaborn==0.13.1
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
supervision==0.17.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.2
urllib3==2.0.7
Werkzeug==3.0.1
yolov5==7.0.13
